// apps/org/src/types/db.d.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface AppEntityLifecycleEvent {
  created_at: Generated<Timestamp>;
  entity_lifecycle_event_id: Generated<string>;
  event: string;
  updated_at: Generated<Timestamp>;
}

export interface AppMemberRole {
  created_at: Generated<Timestamp>;
  member_role_id: Generated<string>;
  role: string;
  updated_at: Generated<Timestamp>;
}

export interface AppOrganization {
  created_at: Generated<Timestamp>;
  creator_user_id: string;
  default_user_organization: boolean;
  deleted_at: Timestamp | null;
  description: string | null;
  name: string;
  organization_id: Generated<string>;
  slug: string;
  updated_at: Generated<Timestamp>;
}

export interface AppOrganizationDescriptionHistory {
  created_at: Generated<Timestamp>;
  description: string | null;
  fk_organization_organization_id: string;
  initiator_user_id: string;
  organization_description_history_id: Generated<string>;
}

export interface AppOrganizationLifecycleEventHistory {
  created_at: Generated<Timestamp | null>;
  entity_lifecycle_event_id: string;
  fk_organization_organization_id: string;
  initiator_user_id: string;
  organization_lifecycle_event_history_id: string;
}

export interface AppOrganizationMemberRole {
  created_at: Generated<Timestamp>;
  organization_member_role_id: Generated<string>;
  role: string;
}

export interface AppOrganizationMembership {
  created_at: Generated<Timestamp>;
  fk_organization_member_role_organization_member_role_id: string;
  fk_organization_organization_id: string;
  member_user_id: string;
  organization_membership_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface AppOrganizationMembershipHistory {
  created_at: Generated<Timestamp>;
  fk_organization_member_role_organization_member_role_id: string;
  fk_organization_organization_id: string;
  initiator_user_id: string;
  member_user_id: string;
  organization_membership_history_id: Generated<string>;
}

export interface AppOrganizationNameHistory {
  created_at: Generated<Timestamp>;
  fk_organization_organization_id: string;
  initiator_user_id: string;
  name: string;
  organization_name_history_id: Generated<string>;
}

export interface AppOrganizationSlugHistory {
  created_at: Generated<Timestamp>;
  fk_organization_organization_id: string;
  initiator_user_id: string;
  organization_slug_history_id: Generated<string>;
  slug: string;
}

export interface AppProject {
  created_at: Generated<Timestamp>;
  creator_user_id: string;
  deleted_at: Timestamp | null;
  description: string;
  fk_team_team_id: string;
  name: string;
  project_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface AppProjectDescriptionHistory {
  created_at: Generated<Timestamp>;
  description: string;
  initiator_user_id: string;
  project_description_history: Generated<string>;
}

export interface AppProjectLifecycleEventHistory {
  created_at: Generated<Timestamp | null>;
  entity_lifecycle_event_id: string;
  fk_project_project_id: string;
  initiator_user_id: string;
  project_lifecycle_event_history_id: string;
}

export interface AppProjectOwnerTeam {
  created_at: Generated<Timestamp>;
  fk_project_project_id: string;
  fk_team_team_id: string;
  project_owner_team_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface AppProjectOwnerTeamHistory {
  created_at: Generated<Timestamp>;
  fk_project_project_id: string;
  fk_team_team_id: string;
  initiator_user_id: string;
  project_owner_team_history_id: Generated<string>;
}

export interface AppTeam {
  created_at: Generated<Timestamp>;
  creator_user_id: string;
  deleted_at: Timestamp | null;
  description: string;
  name: string;
  team_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface AppTeamDescriptionHistory {
  created_at: Generated<Timestamp>;
  description: string;
  initiator_user_id: string;
  team_description_history_id: Generated<string>;
}

export interface AppTeamLifecycleEventHistory {
  created_at: Generated<Timestamp | null>;
  entity_lifecycle_event_id: string;
  fk_team_team_id: string;
  initiator_user_id: string;
  team_lifecycle_event_history_id: string;
}

export interface AppTeamMembership {
  created_at: Generated<Timestamp>;
  fk_member_role_member_role_id: string;
  fk_team_team_id: string;
  member_user_id: string;
  team_membership_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface AppTeamMembershipHistory {
  created_at: Generated<Timestamp>;
  fk_member_role_member_role_id: string;
  fk_team_team_id: string;
  initiator_user_id: string;
  member_user_id: string;
  team_membership_history_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface AppTeamNameHistory {
  created_at: Generated<Timestamp>;
  initiator_user_id: string;
  name: string;
  team_name_history_id: Generated<string>;
}

export interface AppTeamOwnerOrganization {
  created_at: Generated<Timestamp>;
  fk_organization_organization_id: string;
  fk_team_team_id: string;
  team_owner_organization_id: Generated<string>;
  updated_at: Generated<Timestamp>;
}

export interface DB {
  "app.entity_lifecycle_event": AppEntityLifecycleEvent;
  "app.member_role": AppMemberRole;
  "app.organization": AppOrganization;
  "app.organization_description_history": AppOrganizationDescriptionHistory;
  "app.organization_lifecycle_event_history": AppOrganizationLifecycleEventHistory;
  "app.organization_member_role": AppOrganizationMemberRole;
  "app.organization_membership": AppOrganizationMembership;
  "app.organization_membership_history": AppOrganizationMembershipHistory;
  "app.organization_name_history": AppOrganizationNameHistory;
  "app.organization_slug_history": AppOrganizationSlugHistory;
  "app.project": AppProject;
  "app.project_description_history": AppProjectDescriptionHistory;
  "app.project_lifecycle_event_history": AppProjectLifecycleEventHistory;
  "app.project_owner_team": AppProjectOwnerTeam;
  "app.project_owner_team_history": AppProjectOwnerTeamHistory;
  "app.team": AppTeam;
  "app.team_description_history": AppTeamDescriptionHistory;
  "app.team_lifecycle_event_history": AppTeamLifecycleEventHistory;
  "app.team_membership": AppTeamMembership;
  "app.team_membership_history": AppTeamMembershipHistory;
  "app.team_name_history": AppTeamNameHistory;
  "app.team_owner_organization": AppTeamOwnerOrganization;
}